// src/app/Model/listingBuyer.ts
export class listingBuyer {
  key: string;
  Firstname: string;
  Lastname: string;
  Email: string;
  DOB: string;
  Password: string;
  Confirm: string;
  financial: string;
  ChainStatus: string;
  cars: string;
  FinancialPosition: any;
  Position: string = "Cash Buyer";
  CurrentAddress: string;
  PropertyType: string = "Flat";
  Country: string;
  addressname: string;
  Currentstate: string;
  CurrentTown: string;
  Currentpostcode: any;
  Lookingpostcode: any;
  LookingTown: any;
  Lookingstate: any;
  LookingAddress: any;
  latitude: any;
  longitude: any;

  PriceRange: any = 0;
  Validity: any;
  Conditions: any = "any";
  Bcurrent: string;
  Blooking: string;
  MinAmount: any;
  MaxAmount: any;
  LookingStreetname:string;
  phonenumber:number;
  otherInfo:string;
  active = true;
  static title: string;
  static Firstname: string;
  result: any;
  data: any;
  static Lookingpostcode: string;
  UserId : any;
  Roomsmax;
  Roommin:string;
  Ownership:any = "any";
  Minbathroom:any = "any";
  Maxbathroom:any = "any";
  Minreception:any= "any";
  Maxreception:any = "any";
  features:any ="any" ;
  Currentcountry:string;
  username:any;
  useremail:any;
  userdob:any;
  userphone:any;
  usertitle:any;
  title:any;
  PropertyFor:any;
}